app/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
aiomysql==0.2.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
aiofiles==23.2.1
google-cloud-storage==2.16.0
pandas==2.2.2
reportlab==4.1.0
httpx==0.27.0
cryptography==42.0.7
bcrypt==4.0.1
